# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JernihAI API
  description: 'API backend JernihAI — platform peningkatan kualitas gambar.


    Dokumentasi API Publik B2B (FR-14) untuk developer: `docs/API_B2B.md`. Spesifikasi
    OpenAPI lengkap: `docs/api/openapi.yaml` (regenerate via `python api/scripts/export_openapi.py`).'
  version: 0.1.0
paths:
...
  /api/v1/jobs:
    post:
      tags:
      - jobs
      summary: Upload gambar & mulai proses (FR-02/FR-03)
      operationId: create_job_api_v1_jobs_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_job_api_v1_jobs_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - jobs
      summary: Riwayat job user (FR-10)
      description: 'Daftar job milik user, terbaru dulu (FR-10).


        Hanya job milik user yang login — tidak membocorkan job orang lain.

        Pagination `limit`/`offset`; `result_deleted_at` memberitahu UI bahwa

        hasil sudah dihapus retensi (tombol unduh ulang dinonaktifkan).'
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/batch:
    post:
      tags:
      - jobs
      summary: Upload beberapa gambar sekaligus (FR-12)
      description: 'Buat hingga `batch_max_files` job dalam satu request (FR-12).


        Validasi SEMUA file dulu (sebelum simpan apa pun) agar request gagal

        atomik: satu file tidak valid -> tidak ada job yang dibuat, kuota

        tidak terpotong. Kuota dicek untuk total batch (FR-06).'
      operationId: create_batch_jobs_api_v1_jobs_batch_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_batch_jobs_api_v1_jobs_batch_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Cek status job (polling — FR-03)
      operationId: get_job_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}/download:
    get:
      tags:
      - jobs
      summary: Unduh hasil proses (FR-05)
      operationId: download_result_api_v1_jobs__job_id__download_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/jobs:
    get:
      tags:
      - admin
      summary: Semua job lintas user (FR-13)
      description: 'Daftar job (terbaru dulu) — semua user atau riwayat SATU user
        (`email`).


        `email` = pencocokan persis (email unik) — dipakai admin melihat

        riwayat lengkap milik user tertentu.'
      operationId: admin_jobs_api_v1_admin_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminJobList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/jobs/{job_id}:
    delete:
      tags:
      - admin
      summary: Hapus job + file-nya (alat admin)
      description: 'Hapus permanen job (baris DB + file original & hasil di storage).


        File hanya dihapus bila berada di dalam upload_dir/result_dir (guard

        path traversal via `delete_if_inside`; berlaku untuk backend lokal

        maupun R2). Dipakai pengelola membersihkan data uji coba.'
      operationId: admin_delete_job_api_v1_admin_jobs__job_id__delete
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Admin Delete Job Api V1 Admin Jobs  Job Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/users/{user_id}/jobs:
    delete:
      tags:
      - admin
      summary: Hapus SEMUA job user + file-nya (alat admin)
      description: 'Hapus permanen SEMUA job milik satu user (baris DB + file di disk).


        Dipakai halaman detail admin untuk membersihkan data uji coba satu

        user sekaligus — job milik user lain tidak tersentuh.'
      operationId: admin_delete_user_jobs_api_v1_admin_users__user_id__jobs_delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Admin Delete User Jobs Api V1 Admin Users  User Id  Jobs
                  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/b2b/jobs:
    post:
      tags:
      - b2b
      summary: 'B2B: upload gambar & mulai proses (FR-14)'
      description: 'Pay-per-call: 1 job = 1 kredit dari saldo pemilik key.


        Potongan kredit bersifat **atomik** (2 request bersamaan tidak bisa

        memakai saldo yang sama). Tanpa saldo -> 402 Payment Required.

        Status awal `queued` — job diproses worker asinkron; polling dengan

        `GET /b2b/jobs/{id}` hingga `completed`.'
      operationId: b2b_create_job_api_v1_b2b_jobs_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_b2b_create_job_api_v1_b2b_jobs_post'
        required: true
      responses:
        '201':
          description: Job dibuat (status awal `queued`) dan 1 kredit terpotong dari
            saldo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '400':
          description: '`scale`/`output_format` tidak dikenal, atau PNG melebihi batas
            4096 px sisi terpanjang (ADR-004)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: scale harus salah satu dari [2, 4]
        '402':
          description: Saldo kredit kosong — isi lewat halaman Billing web
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: 'Saldo kredit kosong (B2B: 1 gambar = 1 kredit). Isi saldo
                  via halaman Billing web.'
        '413':
          description: File melebihi batas ukuran (10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Ukuran maksimal 10 MB
        '415':
          description: Format file tidak didukung — hanya JPG/PNG/WebP (dicek konten)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Hanya menerima JPG, PNG, atau WebP (validasi konten, bukan
                  ekstensi)
        '422':
          description: Validasi form gagal (mis. `scale` bukan int, `file` kosong)
            — `detail` berupa LIST error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
              example:
                detail:
                - loc:
                  - body
                  - name
                  msg: Field required
                  type: value_error
        '401':
          description: Header `X-API-Key` tidak ada atau key tidak dikenal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Header X-API-Key wajib diisi (lihat /api-keys)
        '403':
          description: Key dinonaktifkan (dicabut) atau akun pemilik di-suspend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: API key dinonaktifkan (dicabut oleh pemilik)
        '429':
          description: Rate limit per menit tier terlampaui (free 20 / pro 120)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Terlalu banyak permintaan — coba lagi beberapa saat lagi.
  /api/v1/b2b/jobs/{job_id}:
    get:
      tags:
      - b2b
      summary: 'B2B: cek status job (polling — FR-14)'
      description: Status job milik pemilik key — job key lain = 404 (tanpa bocor
        info).
      operationId: b2b_get_job_api_v1_b2b_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Status job terkini (`queued`/`processing`/`completed`/`failed`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '404':
          description: Job tidak ditemukan (atau milik key lain — tanpa bocor info)
          content:
            application/json:
              example:
                detail: Job tidak ditemukan
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Header `X-API-Key` tidak ada atau key tidak dikenal
          content:
            application/json:
              example:
                detail: Header X-API-Key wajib diisi (lihat /api-keys)
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Key dinonaktifkan (dicabut) atau akun pemilik di-suspend
          content:
            application/json:
              example:
                detail: API key dinonaktifkan (dicabut oleh pemilik)
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit per menit tier terlampaui (free 20 / pro 120)
          content:
            application/json:
              example:
                detail: Terlalu banyak permintaan — coba lagi beberapa saat lagi.
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminJobList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminJobOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: AdminJobList
...
    Body_b2b_create_job_api_v1_b2b_jobs_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
          description: 'Gambar input: JPG/PNG/WebP, maks 10 MB (validasi magic bytes,
            bukan ekstensi).'
        scale:
          type: integer
          title: Scale
          description: Faktor pembesaran (2x atau 4x).
          default: 2
          examples:
          - 2
          - 4
        output_format:
          type: string
          title: Output Format
          description: 'Format hasil: `webp` (default, terbaik), `jpeg`, atau `png`
            (lossless — dibatasi ≤ 4096 px sisi terpanjang, ADR-004).'
          default: webp
          examples:
          - webp
          - jpeg
          - png
        face_enhance:
          type: boolean
          title: Face Enhance
          description: Restorasi wajah (GFPGAN) — memperjelas wajah pada foto lama.
          default: false
        denoise:
          type: boolean
          title: Denoise
          description: Kurangi noise/grain pada foto lama.
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          description: Pertegas warna (saturasi/kontras) untuk foto pudar.
          default: false
      type: object
      required:
      - file
      title: Body_b2b_create_job_api_v1_b2b_jobs_post
    Body_create_batch_jobs_api_v1_jobs_batch_post:
      properties:
        files:
          items:
            type: string
            contentMediaType: application/octet-stream
          type: array
          title: Files
        scale:
          type: integer
          title: Scale
          default: 2
        output_format:
          type: string
          title: Output Format
          default: webp
        face_enhance:
          type: boolean
          title: Face Enhance
          default: false
        denoise:
          type: boolean
          title: Denoise
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          default: false
      type: object
      required:
      - files
      title: Body_create_batch_jobs_api_v1_jobs_batch_post
    Body_create_job_api_v1_jobs_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        scale:
          type: integer
          title: Scale
          default: 2
        output_format:
          type: string
          title: Output Format
          default: webp
        face_enhance:
          type: boolean
          title: Face Enhance
          default: false
        denoise:
          type: boolean
          title: Denoise
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          default: false
      type: object
      required:
      - file
      title: Body_create_job_api_v1_jobs_post
...
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
      description: 'Bentuk tubuh respons error — FastAPI HTTPException selalu `{"detail":
        ...}`.'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobListOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/JobOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: JobListOut
      description: Halaman riwayat job user (FR-10).
    JobOut:
      properties:
        id:
          type: string
          title: Id
        status:
          type: string
          title: Status
        scale:
          type: integer
          title: Scale
        output_format:
          type: string
          title: Output Format
        face_enhance:
          type: boolean
          title: Face Enhance
          default: false
        denoise:
          type: boolean
          title: Denoise
          default: false
        color_enhance:
          type: boolean
          title: Color Enhance
          default: false
        original_name:
          type: string
          title: Original Name
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        created_at:
          type: string
          format: date-time
          title: Created At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        result_deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Result Deleted At
      type: object
      required:
      - id
      - status
      - scale
      - output_format
      - original_name
      - error
      - created_at
      - finished_at
      title: JobOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...